fix: migration scripts
