+++ apps/backend/src/issuer/configuration/credentials/utils/index.ts
@@ -5,5 +5,3 @@ export {
     buildDisclosureFrame,
     buildJsonSchema,
 } from "./derive";
-export { convertV1ToV2, deriveRuntimeArtifacts } from "./migrate";
-export type { CredentialConfigV1 } from "./v2-types";
